+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 - [+] #7 add option to skip feed rendering
 - [\*] #8 render mainpage only once in case mainpage content is in pages
+- [+] #2 pagination for posts (mainpage) list
 
 
 ## 0.2.0 (2016-10-05)